+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Edit: Load Balancer, more restrictive TLS/ciphers, cross-zone load balancing, CloudWatch alarms
 - Edit: WAF: rate limits, SQLi/Unix/Linux rule sets, Geo Blocking
 - Edit: SSM now uses its own KMS key
+- Edit: Bucket Private Logging Bucket, Allow CloudWatch Logs (BucketOwnerPreferred)
 - Add: VPC Endpoints to modules/vpc
 - Add: Module for AMI Lookup
 - Add: Enable switches for IGW/NAT to modules/vpc (default: false)